function to read data from relational databases hosted by online MS SQL server. The user needs to have read access to the database et should install the appropriate MS driver that is compatible with the SQLServer version

function to read data from relational databases hosted by online MS SQL server. The user needs to have read access to the database et should install the appropriate MS driver that is compatible with the SQLServer version

Usage

read_from_ms_sql_server(
  user,
  password,
  host,
  port,
  database.name,
  driver.name,
  table.name,
  records = NULL,
  fields = NULL,
  id.position = 1
)

Arguments

user

the user name

password

the user password

host

the name of the host server

port

the port ID

database.name

the name of the database that contains the table from which the data should be pulled

driver.name

the name of the MS driver. use odbc::odbcListDrivers() to display the list of installed drivers

table.name

the name of the target table

records

a vector or a comma-separated string of subset of subject IDs. When specified, only the records that correspond to these subjects will be imported.

fields

a vector or a comma-separated string of column names. If provided, only those columns will be imported

id.position

the column position of the variable that unique identifies the subjects. This should only be specified when the column with the subject IDs is not the first column. default is 1.

Value

a data frame

Examples

data = read_from_ms_sql_server(user="kmane", password="Dakabantang@KD23", host="robin.mrc.gm", port=1433, database.name="IBS_BHDSS", table.name="dss_events", driver.name="ODBC Driver 17 for SQL Server")