function to read data from REDCap

function to read data from REDCap

Usage

read_from_redcap(
  uri,
  token,
  project.id,
  id.position = 1L,
  records = NULL,
  fields = NULL
)

Arguments

uri

the URI of the server

token

the user-specific string that serves as the password for a project

project.id

the name of the project you want to import data from

id.position

the column position of the variable that unique identifies the subjects. This should only be specified when the column with the subject IDs is not the first column. default is 1.

records

a vector or a comma-separated string of subset of subject IDs. When specified, only the records that correspond to these subjects will be imported.

fields

a vector or a comma-separated string of column names. If provided, only those columns will be imported.

Value

a list with 2 data frames: the data of interest and the metadata associated to the data.

Examples

redcap.data = read_from_redcap(uri="https://redcap.mrc.gm:8443/redcap/api/", token="9D71857D60F4016AB7BFFDA65970D737", project.id="Pats__Covid_19_Cohort_1_Screening", id.position=1, records=NULL, fields=NULL)
#> Starting to read 280 records  at 2023-01-16 15:40:04.
#> 
#> ── Column specification ────────────────────────────────────────────────────────
#> cols(
#>   .default = col_double(),
#>   day_1_q_ran_id = col_character(),
#>   redcap_event_name = col_character(),
#>   day_1_q_1a = col_character(),
#>   day_1_q_1b = col_character(),
#>   day_1_q_1c = col_character(),
#>   day_1_q_1 = col_character(),
#>   day_1_q_2 = col_datetime(format = ""),
#>   day_1_q_4 = col_character(),
#>   day_1_q_6c = col_character(),
#>   day_1_oth_spc_vac = col_logical(),
#>   day_1_q_9e = col_date(format = ""),
#>   day_1_9hi = col_character(),
#>   day_1_q_9f_2 = col_date(format = ""),
#>   day_1_q_11 = col_time(format = ""),
#>   day_1_q_14 = col_character(),
#>   day_1_q_16 = col_character(),
#>   day_1_q_16a = col_character(),
#>   day_1_q_16d = col_character(),
#>   day_1_q_17 = col_character(),
#>   day_1_q_17a = col_character()
#>   # ... with 509 more columns
#> )
#> ℹ Use `spec()` for the full column specifications.