Function to import epidemiology-related data from different sources into R

the function allows import of data stored in database management systems (DBMS) including both relational and NoSQL databases. It also contains functions to import data from common file types such as csv, txt, xlsx, xml, json, etc.

Usage

readepi(
  credentials.file = NULL,
  file.path = NULL,
  sep = NULL,
  format = NULL,
  which = NULL,
  pattern = NULL,
  project.id = NULL,
  driver.name = NULL,
  table.name = NULL,
  records = NULL,
  fields = NULL,
  id.position = 1
)

Arguments

credentials.file

the path to the file with the user-specific credential details for the projects of interest. It is required when importing data from DBMS. This is a tab-delimited file with the following columns:

user_name: the user name

password: the user password (for REDCap, this corresponds to the token that serves as password to the project)

host_name: the host name (for MS SQL servers) or the URI (for REDCap)

project_id: the project ID or the name of the database you are access to.

comment: a summary description about the project or database of interest

dbms: the name of the DBMS: 'redcap' or 'REDCap' when reading from REDCap, 'sqlserver' or 'SQLServer' when reading from MS SQLServer

port: the port ID (used for MS SQLServers)

file.path

the path to the file to be read. When several files need to be imported from a directory, this should be the path to that directory

sep

the separator between the columns in the file. This is only required for space-separated files.

format

a string used to specify the file format. This is useful when a file does not have an extension, or has a file extension that does not match its actual type

which

a string used to specify which objects should be extracted (the name of the excel sheet to import)

pattern

when specified, only files with this suffix will be imported from the specified directory

project.id

for relational DB, this is the name of the database that contains the table from which the data should be pulled. Otherwise, it is the project ID you were given access to.

driver.name

the name of the MS driver. use odbc::odbcListDrivers() to display the list of installed drivers

table.name

the name of the target table

records

a vector or a comma-separated string of subset of subject IDs. When specified, only the records that correspond to these subjects will be imported.

fields

a vector or a comma-separated string of column names. If provided, only those columns will be imported.

id.position

the column position of the variable that unique identifies the subjects. This should only be specified when the column with the subject IDs is not the first column. default is 1.

Value

a list with 2 data frames (data and metadata) when reading from REDCap. A data frame otherwise.

Examples

data = readepi(credentials.file=system.file("extdata", "test.ini", package = "readepi"), project.id="Pats__Covid_19_Cohort_1_Screening")
#> Starting to read 280 records  at 2023-01-16 15:40:18.
#> 
#> ── Column specification ────────────────────────────────────────────────────────
#> cols(
#>   .default = col_double(),
#>   day_1_q_ran_id = col_character(),
#>   redcap_event_name = col_character(),
#>   day_1_q_1a = col_character(),
#>   day_1_q_1b = col_character(),
#>   day_1_q_1c = col_character(),
#>   day_1_q_1 = col_character(),
#>   day_1_q_2 = col_datetime(format = ""),
#>   day_1_q_4 = col_character(),
#>   day_1_q_6c = col_character(),
#>   day_1_oth_spc_vac = col_logical(),
#>   day_1_q_9e = col_date(format = ""),
#>   day_1_9hi = col_character(),
#>   day_1_q_9f_2 = col_date(format = ""),
#>   day_1_q_11 = col_time(format = ""),
#>   day_1_q_14 = col_character(),
#>   day_1_q_16 = col_character(),
#>   day_1_q_16a = col_character(),
#>   day_1_q_16d = col_character(),
#>   day_1_q_17 = col_character(),
#>   day_1_q_17a = col_character()
#>   # ... with 509 more columns
#> )
#> ℹ Use `spec()` for the full column specifications.