Import data from REDCap into R

Import data from REDCap into R

Usage

read_from_redcap(
  uri,
  token,
  id_position = NULL,
  id_col_name = NULL,
  records = NULL,
  fields = NULL
)

Arguments

uri

the URI of the server

token

the user-specific string that serves as the password for a project

id_position

the column position of the variable that unique identifies the subjects. When the column name with the subject IDs is known, use the id_col_name argument instead. default is 1

id_col_name

the column name with the subject IDs

records

a vector or a comma-separated string of subset of subject IDs. When specified, only the records that correspond to these subjects will be imported.

fields

a vector or a comma-separated string of column names. If provided, only those columns will be imported.

Value

a list with 2 data frames: the data of interest and the metadata associated to the data.

Examples

redcap_data <- read_from_redcap(
  uri = "https://bbmc.ouhsc.edu/redcap/api/",
  token = "9A81268476645C4E5F03428B8AC3AA7B",
  id_position = 1,
  id_col_name = NULL,
  records = NULL,
  fields = NULL
)
#> Starting to read 5 records  at 2023-05-05 10:01:33.
#> 
#> ── Column specification ────────────────────────────────────────────────────────
#> cols(
#>   .default = col_double(),
#>   name_first = col_character(),
#>   name_last = col_character(),
#>   address = col_character(),
#>   telephone = col_character(),
#>   email = col_character(),
#>   dob = col_date(format = ""),
#>   comments = col_character(),
#>   mugshot = col_character()
#> )
#> ℹ Use `spec()` for the full column specifications.